Cleanup buildSrc and gradle files:
- Remove unused github-publish plugin
- Remove unnecessary publishAsDraft variable
